chore: rename project as `thumbs` and add LICENSE
- The whole project is renamed as `thumbs` and `thumbs-android`.
- MIT LICENSE is added.

Fixes https://github.com/thumbs-project/thumbs-project/issues/18
