+++ README.md
@@ -1,2 +1,2 @@
-# tamagotchi-android
-tamagotchi-android
+# thumbs-android
+**Thumbs Android** is a client-side (on Android) application of thumbs-project. The application runs with the Android Software Development Kit (SDK) of minimum version 23.
