+++ app/build.gradle
@@ -7,11 +7,11 @@ apply plugin: 'kotlin-android-extensions'
 android {
   compileSdkVersion 28
   defaultConfig {
-    applicationId "com.tamagotchi.android.tamagotchi_android"
+    applicationId "com.thumbs.android.thumbs_android"
     minSdkVersion 23
     targetSdkVersion 28
     versionCode 1
-    versionName "1.0"
+    versionName "0.0.1"
     testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
   }
   buildTypes {
