--- app/src/main/java/com/tamagotchi/android/tamagotchiAndroid/modules/domain/model/User.kt
@@ -1,8 +1,0 @@
-package com.tamagotchi.android.tamagotchiAndroid.modules.domain.model
-
-
-
-data class User(
-  var id : String,
-  var url : String
-)
