--- app/src/main/java/com/tamagotchi/android/tamagotchiAndroid/modules/domain/repositories/UserRepository.kt
@@ -1,9 +1,0 @@
-package com.tamagotchi.android.tamagotchiAndroid.modules.domain.repositories
-
-import com.tamagotchi.android.tamagotchiAndroid.modules.domain.model.User
-import io.reactivex.Single
-
-
-interface UserRepository{
-  fun getUser() : Single<User>
-}
