+++ app/src/main/java/com/thumbs/android/thumbsAndroid/modules/domain/model/User.kt
@@ -1,0 +1,8 @@
+package com.thumbs.android.thumbsAndroid.modules.domain.model
+
+
+
+data class User(
+  var id : String,
+  var url : String
+)
