+++ app/src/main/java/com/thumbs/android/thumbsAndroid/modules/domain/repositories/UserRepository.kt
@@ -1,0 +1,9 @@
+package com.thumbs.android.thumbsAndroid.modules.domain.repositories
+
+import com.thumbs.android.thumbsAndroid.modules.domain.model.User
+import io.reactivex.Single
+
+
+interface UserRepository{
+  fun getUser() : Single<User>
+}
